databasesettings implemented

+++ php/profiler/config/deleterow.php
@@ -2,27 +2,31 @@
 <?php
 
 	include 'print.php';
+	include '../loadconnection.php';
 
 	$table = $_GET["tablename"];
 	$row = $_GET["row"];
 	
 	if(strlen($table)>0 and strlen($row)>0){
 	
-		$con = mysqli_connect("localhost","testuser","testpw","configtest");
-		if (mysqli_connect_errno())
+		if($connection = loadConnection())
 		{
-			echo "Failed to connect to MySQL: " . mysqli_connect_error();
-		}
+			$con = mysqli_connect("localhost",$connection[1],$connection[2],$connection[0]);
+			if (mysqli_connect_errno())
+			{
+				echo "Failed to connect to MySQL: " . mysqli_connect_error();
+			}
 
-		$sql = "DELETE FROM ".$table." WHERE category = '$row'";
+			$sql = "DELETE FROM ".$table." WHERE category = '$row'";
 
-		if (mysqli_query($con,$sql))
-		{
-			echo printTables($con);
-		}
-		else
-		{
-			echo "ERROR: " . mysqli_error($con);
+			if (mysqli_query($con,$sql))
+			{
+				echo printTables($con);
+			}
+			else
+			{
+				echo "ERROR: " . mysqli_error($con);
+			}
 		}
 	}
 ?>
