+++ php/profiler/config/deletetable.php
@@ -1,33 +1,37 @@
 <?php
 
 	include 'print.php';
+	include '../loadconnection.php';
 
 	$table = $_GET["tablename"];
 	
 	if(strlen($table)>0){
 	
-		$sql = "DROP TABLE $table";
-		
-		$con = mysqli_connect("localhost","testuser","testpw","configtest");
-		if (mysqli_connect_errno())
+		if($connection = loadConnection())
 		{
-			echo "Failed to connect to MySQL: " . mysqli_connect_error();
-		}
-
-		if (mysqli_query($con,$sql))
-		{
-			if (mysqli_query($con,"ALTER TABLE uservalues DROP history_".$table))
+			$sql = "DROP TABLE $table";
+			
+			$con = mysqli_connect("localhost",$connection[1],$connection[2],$connection[0]);
+			if (mysqli_connect_errno())
 			{
-				echo printTables($con);
+				echo "Failed to connect to MySQL: " . mysqli_connect_error();
+			}
+
+			if (mysqli_query($con,$sql))
+			{
+				if (mysqli_query($con,"ALTER TABLE uservalues DROP history_".$table ))
+				{
+					echo printTables($con);
+				}
+				else
+				{
+					echo "Error creating table: " . mysqli_error($con);
+				}
 			}
 			else
 			{
 				echo "Error creating table: " . mysqli_error($con);
 			}
-		}
-		else
-		{
-			echo "Error creating table: " . mysqli_error($con);
 		}
 	}
 ?>
